+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This changelog contains a loose collection of changes in every release. I will a
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## Unreleased
+
+### Fixed
+
+* Parsing of single-statement policies when expressed as a JSON object instead of a JSON object wrapped into an array (#10)
+
 ## 1.3.0 - 2020-01-26
 
 ### Added